Refactored file locations for clarity
