+++ Tests/newspeedtest.py
