+++ shapeshifter/__init__.py
@@ -1,8 +1,8 @@
 __all__ = ['ColumnInfo', 'ShapeShifter', 'OperatorEnum', 'ContinuousQuery', 'DiscreteQuery', 'FileTypeEnum']
 
-from shapeshifter.columninfo import ColumnInfo
 from shapeshifter.shapeshifter import ShapeShifter
-from shapeshifter.operatorenum import OperatorEnum
-from shapeshifter.continuousquery import ContinuousQuery
-from shapeshifter.discretequery import DiscreteQuery
-from shapeshifter.filetypeenum import FileTypeEnum
+from shapeshifter.utils.columninfo import ColumnInfo
+from shapeshifter.utils.continuousquery import ContinuousQuery
+from shapeshifter.utils.discretequery import DiscreteQuery
+from shapeshifter.utils.filetypeenum import FileTypeEnum
+from shapeshifter.utils.operatorenum import OperatorEnum
