+++ shapeshifter/utils/columninfo.py
