+++ shapeshifter/utils/continuousquery.py
@@ -1,9 +1,11 @@
-from shapeshifter import operatorenum
+from shapeshifter.utils import operatorenum
+
+
 class ContinuousQuery:
 	"""
 	Data object that encapsulates a continuous query, holding info on the column, operator, and value.
 	"""
-	def __init__(self, columnName:str, operator:operatorenum, value:float):
+	def __init__(self, columnName:str, operator: operatorenum, value:float):
 		"""
 		:type columnName: string
 		:param columnName: name of the column to be examined
