+++ shapeshifter/utils/discretequery.py
