+++ shapeshifter/utils/filetypeenum.py
