+++ shapeshifter/utils/operatorenum.py
